- add item category filter
- add item category as filter (#11, #39)
- add prophecies item type base (#37)
- fix mods without values (#39)
