+++ src/app/shared/module/poe/provider/base-item-type-categories.provider.ts
@@ -1,0 +1,12 @@
+import { Injectable } from '@angular/core';
+import { Default } from '../../../../../assets/poe/base-item-type-categories.json';
+import { BaseItemTypeCategoryMap } from '../type';
+
+@Injectable({
+    providedIn: 'root'
+})
+export class BaseItemTypeCategoriesProvider {
+    public provide(): BaseItemTypeCategoryMap {
+        return Default as BaseItemTypeCategoryMap;
+    }
+}
