+++ src/app/shared/module/poe/provider/index.ts
@@ -1,4 +1,4 @@
-export * from './base-item-type.provider';
+export * from './base-item-types.provider';
 export * from './client-string.provider';
 export * from './currency/currencies.provider';
 export * from './currency/currency-chaos-equivalents.provider';
