+++ src/app/shared/module/poe/service/index.ts
@@ -1,8 +1,6 @@
-export * from './base-item-type/base-item-type.service';
 export * from './context.service';
 export * from './item/item-search-evaluate.service';
 export * from './item/item-search.service';
 export * from './item/parser/item-parser.service';
 export * from './stats-description/stats-description.service';
 export * from './word/word.service';
-
