+++ src/app/shared/module/poe/service/item/item-search.service.spec.ts
@@ -1,14 +1,14 @@
 import { async, TestBed } from '@angular/core/testing';
 import { Item, Language } from '@shared/module/poe/type';
 import { SharedModule } from '@shared/shared.module';
-import { BaseItemTypeService } from '../base-item-type/base-item-type.service';
+import { BaseItemTypesService } from '../base-item-types/base-item-types.service';
 import { ContextService } from '../context.service';
 import { ItemSearchService } from './item-search.service';
 
 describe('ItemSearchService', () => {
     let sut: ItemSearchService;
     let contextService: ContextService;
-    let baseItemTypeService: BaseItemTypeService;
+    let baseItemTypesService: BaseItemTypesService;
 
     beforeEach(async(() => {
         TestBed.configureTestingModule({
@@ -22,12 +22,12 @@ describe('ItemSearchService', () => {
         contextService.init({
             language: Language.English
         });
-        baseItemTypeService = TestBed.get<BaseItemTypeService>(BaseItemTypeService);
+        baseItemTypesService = TestBed.get<BaseItemTypesService>(BaseItemTypesService);
     }));
 
     it('should return items', (done) => {
         const requestedItem: Item = {
-            typeId: baseItemTypeService.search('Topaz Ring')
+            typeId: baseItemTypesService.search('Topaz Ring')
         };
 
         sut.search(requestedItem).subscribe(result => {
