+++ src/app/shared/module/poe/service/item/parser/item-section-explicits-parser.service.ts
@@ -23,13 +23,7 @@ export class ItemSectionExplicitsParserService implements ItemSectionParserServi
                 return null;
         }
 
-        const result: Section[] = [];
-        for (const section of item.sections) {
-            if (this.parseSection(section, target)) {
-                result.push(section);
-            }
-        }
-        return result;
+        return item.sections.filter(section => this.parseSection(section, target));
     }
 
     private parseSection(section: Section, target: Item): boolean {
