+++ src/assets/poe/words.json
@@ -1090,7 +1090,7 @@
 		"1088": "L'Étreinte d'une mère",
 		"1089": "Héritage du Guerrier",
 		"1090": "Projecteur astral",
-		"1091": "Le Sauveur",
+		"1091": "La Salvatrice",
 		"1092": "Don du saule",
 		"1093": "Lueur d'espoir",
 		"1094": "Œil de la Malveillance",
@@ -2189,7 +2189,7 @@
 		"L'Étreinte d'une mère": "1088",
 		"Héritage du Guerrier": "1089",
 		"Projecteur astral": "1090",
-		"Le Sauveur": "1091",
+		"La Salvatrice": "1091",
 		"Don du saule": "1092",
 		"Lueur d'espoir": "1093",
 		"Œil de la Malveillance": "1094",
